WIP: Versioning objects, table definitions and store/loads not in heap
This commit is far too broad and captures a bunch of changes that
include comments to myself, minor edits and thoughts.
Mostly this builds tables for versioning objects, and hooks into
the zgc load barier on load/store not in heap, and the load_preloaded,
i'm committing now as small brokes execute to a point and then break
which is interesting!