Can run simple programs without versioning
A lot of the existing infrastructure now exists for building snapshots from commits.
Currently on oop_store_not_in_heap, we create an object number for the object to be stored,
create a version for that object and store the object in our version table,
encode a new zpointer to map to that object pointer to the object number,
record the version number for the new object number, and return the funny zpointer.

When we hit funny zpointers in load_not_in_heap, and load_barrier_on_oop_field_preloaded
we find the object number, find the latest version in the version table, copy the object
into a fresh place in memory, and then return the original verison. This is very wrong,
we should find the right version for this thread, we should return the pointer to the
copied object, and we should map the object we have created an object for this funny
pointer already in this concurrent block, otherwise we may encounter the same
funny pointer for the object number, and create a new object.

I wanted to record a spot in the commit history where programs still seemed to run
correctly wrt the current JVM implementation before i break things.